Darktopia - FATTY UPDATE

+++ code/sepulcher/objects/structures/lamp_post.dm
@@ -5,14 +5,15 @@
 	icon = 'icons/sepulcher/lamp_post.dmi'
 	icon_state = "post"
 	layer = FLY_LAYER
-	var/lantern_having = FALSE
+	var/obj/item/flashlight/lantern/lantern
 
 /obj/structure/lamp_post/Initialize()
 	. = ..()
 	update_brightness()
 
 /obj/structure/lamp_post/proc/update_brightness(mob/user = null)
-	if(lantern_having)
+	if(lantern)
+		icon_state = "post_lantern"
 		set_light(6,1,"#E9CF48") // Abit more range than a freestanding lantern
 	else
 		icon_state = initial(icon_state)
@@ -20,21 +21,22 @@
 
 /obj/structure/lamp_post/attackby(obj/item/L, mob/user, params)
 	if(/obj/item/flashlight/lantern)
-		to_chat(user, "You set the lantern on the hook.")
-		icon_state = "post_lantern"
-		qdel(L)
-		lantern_having = TRUE
+		if(!user.transferItemToLoc(L, src))
+			return
+		lantern = L
 		update_brightness()
+		to_chat(user, "You set the lantern on the hook.")
+		qdel(L)
 
 /obj/structure/lamp_post/attack_hand(mob/user)
-	if(lantern_having)
-		var/obj/item/flashlight/lantern/L = new /obj/item/flashlight/lantern
-		user.put_in_active_hand(L)
+	if(lantern)
+		user.put_in_hands(lantern)
+		lantern = null
 		to_chat(user, "You take the lantern off the hook.")
-		lantern_having = FALSE
 		update_brightness()
 	else
 		return
 
+//Has lantern at start
 /obj/structure/lamp_post/lantern
-	lantern_having = TRUE
+	lantern = new /obj/item/flashlight/lantern
